+++ src/interfaces/positions-contract.ts
@@ -5,7 +5,6 @@ export interface EosdtPosition {
     governance: string // EOS type: asset
     collateral: string // EOS type: float64
 }
-
 export const positionKeys = [
     "position_id",
     "maker",
@@ -19,37 +18,13 @@ export interface Referral {
     referral: string // EOS type: name
     staked_amount: string // EOS type: asset
 }
-
 export const referralKeys = ["referral_id", "referral", "staked_amount"]
 
 export interface PositionReferral {
     referral_id: number // EOS type: uint64
     position_id: number // EOS type: uint64
 }
-
 export const positionReferralKeys = ["referral_id", "position_id"]
-
-export interface TokenRate_deprecated {
-    rate: string // EOS type: asset
-    update: string // EOS type: time_point_sec
-    provablecb1a_price: string // EOS type: asset
-    provablecb1a_update: string // EOS type: time_point_sec
-    delphioracle_price: string // EOS type: asset
-    delphioracle_update: string // EOS type: time_point_sec
-    equilibriumdsp_price: string // EOS type: asset
-    equilibriumdsp_update: string // EOS type: time_point_sec
-}
-
-export const tokenRateKeys_deprecated = [
-    "rate",
-    "update",
-    "provablecb1a_price",
-    "provablecb1a_update",
-    "delphioracle_price",
-    "delphioracle_update",
-    "equilibriumdsp_price",
-    "equilibriumdsp_update"
-]
 
 export interface TokenRate {
     rate: string // EOS type: asset
@@ -60,10 +35,9 @@ export interface TokenRate {
     delphioracle_update: string // EOS type: time_point_sec
     equilibriumdsp_price: string // EOS type: asset
     equilibriumdsp_update: string // EOS type: time_point_sec
-    id : number // EOS type: uint64
+    id: number // EOS type: uint64
     base: string // EOS type: symbol
 }
-
 export const tokenRateKeys = [
     "rate",
     "update",
@@ -81,7 +55,6 @@ export interface LtvRatios {
     position_id: number // EOS type: uint64
     ltv_ratio: string // EOS type: float64
 }
-
 export const ltvRatiosKeys = ["position_id", "ltv_ratio"]
 
 export interface EosdtContractSettings {
@@ -116,8 +89,8 @@ export interface EosdtContractSettings {
 
     collateral_account: string // EOS type: name
     collateral_token: string // EOS type: symbol
+    savings_account: string // EOS type: name
 }
-
 export const contractSettingsKeys = [
     "setting_id",
     "global_lock",
@@ -144,7 +117,8 @@ export const contractSettingsKeys = [
     "referral_min_stake",
     "referral_ratio",
     "collateral_account",
-    "collateral_token"
+    "collateral_token",
+    "savings_account"
 ]
 
 export interface EosdtContractParameters {
@@ -158,8 +132,7 @@ export interface EosdtContractParameters {
     prev_stake: string // EOS type: time_point_sec
     eos_staked: string // EOS type: asset
 }
-
-export const сontractParametersKeys = [
+export const contractParametersKeys = [
     "parameter_id",
     "total_collateral",
     "total_debt",